+++ src/Crowdin.Api/AI/GenerateAiPromptFineTuningDatasetRequest.cs
@@ -11,7 +11,7 @@ namespace Crowdin.Api.AI
     public class GenerateAiPromptFineTuningDatasetRequest
     {
         [JsonProperty("projectIds")]
-        public ICollection<int> ProjectIds { get; set; }
+        public ICollection<long> ProjectIds { get; set; }
         
         [JsonProperty("purpose")]
         public AiDatasetPurpose? Purpose { get; set; }
@@ -23,12 +23,12 @@ namespace Crowdin.Api.AI
         public DateTimeOffset? DateTo { get; set; }
         
         [JsonProperty("maxFileSize")]
-        public int? MaxFileSize { get; set; }
+        public long? MaxFileSize { get; set; }
         
         [JsonProperty("minExamplesCount")]
-        public int? MinExamplesCount { get; set; }
+        public long? MinExamplesCount { get; set; }
         
         [JsonProperty("maxExamplesCount")]
-        public int? MaxExamplesCount { get; set; }
+        public long? MaxExamplesCount { get; set; }
     }
 }
