+++ src/Crowdin.Api/Dictionaries/IDictionariesApiExecutor.cs
@@ -4,13 +4,15 @@ using System.Threading.Tasks;
 
 using JetBrains.Annotations;
 
+#nullable enable
+
 namespace Crowdin.Api.Dictionaries
 {
     [PublicAPI]
     public interface IDictionariesApiExecutor
     {
-        Task<ResponseList<Dictionary>> ListDictionaries(int projectId, string? languageIds = null);
+        Task<ResponseList<Dictionary>> ListDictionaries(long projectId, string? languageIds = null);
 
-        Task<Dictionary> EditDictionary(int projectId, string languageId, IEnumerable<DictionaryPatch> patches);
+        Task<Dictionary> EditDictionary(long projectId, string languageId, IEnumerable<DictionaryPatch> patches);
     }
 }
